screenshot moved from windowManager
